Beta Test Submit 808MastahMind
This is the culmination of all of the work and research that I put in
over break. I spent hours examining the different codes and websites,
using what I learned to create this code. I did copy the guessing
portion of this code but then I modified it a bit to better suit the
idea I had for my version of Mastermind. However I did do the controls,
lights and variables by myself with only the concepts from the Arduino
website.